Charge Wallets Are Works Fine, Edit Wallet Model and it's Repo Update Func, Code Need To be More Cleaned and may be more Complex cause it's very simple

+++ Kushl_3m3bdo/Models/Wallet.cs
@@ -15,7 +15,7 @@ namespace Kushl_3m3bdo.Models
 		public DateTime PlanSubscriptionStartDate { get; set; }
 		public DateTime WalletCreatedDate { get; set; } = DateTime.UtcNow;
 		public int NumberOfPurchases { get; set; } = 0;
-		public long PriceOfPurchases { get; set; } = 0;
+		public decimal PriceOfPurchases { get; set; } = 0;
 		public double Score { get; set; } = 100.0;
